requirements.txt
numpy
pygame
pytorch
tqdm